# api.yml
openapi: "3.0.0"
info:
  description: "Métodos disponibles API GoodWayPilgrim"
  version: "1.0.0"
  title: "API GoodWayPilgrim"
paths:
  /pub/register:
    post:
      tags:
      - "user"
      summary: "Adds a new user"
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: User name
                example: "Irene"
              surname:
                type: string
                description: User surname
                example: "Sánchez"
              email:
                type: string
                description: User email
                example: "irene@example.com"
              password:
                type: string
                description: User password
                example: "Password1"
      responses:
        "200":
          description: "New user created"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: User id
                    example: 1
                  name:
                    type: string
                    description: User name
                    example: "Irene"
                  surname:
                    type: string
                    description: User surname
                    example: "Sánchez"
                  email:
                    type: string
                    description: User email
                    example: "irene@example.com"
                  password:
                    type: string
                    description: User password
                    example: "Password1"
                example: {'id': '1','name': 'Irene','surname': 'Sánchez', 'email': 'irene@gmail.com', 'password': 'Password1'}
        "400":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"Incorrect data recived"}
        "422":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"email is already in database"}
  /pub/login:
    post:
      tags:
      - "user"
      summary: "Log in GoodWayPilgrim"
      parameters:
        - name: identification
          in: query
          required: true
          schema:
            type: object
            properties:
              email:
                type: string
                description: User email
                example: "irene@example.com"
              password:
                type: string
                description: User password
                example: "Password1"
      responses:
        "200":
          description: "Email and password are correct"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: message
                    example: "success"
                  token:
                    type: string
                    description: personal token
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                example: {'message': 'success','token': 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c'}
        "400":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"Incorrect data recived"}
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"email or password is wrong"}
...
  /pri/editProfile:
    put:
      tags:
      - "user"
      summary: "Edits the personal profile"
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: User name
                example: "Irene"
              surname:
                type: string
                description: User surname
                example: "Sánchez"
              email:
                type: string
                description: User email
                example: "irene@example.com"
              oldPassword:
                type: string
                description: Current password
                example: "Password1"
              newPassword:
                type: string
                description: New password
                example: "passUser"
              picture:
                type: string
                description: New profile picture
                example: ""
      responses:
        "200":
          description: "Shows profile"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: User id
                    example: 1
                  name:
                    type: string
                    description: User name
                    example: "Irene"
                  surname:
                    type: string
                    description: User surname
                    example: "Sánchez"
                  email:
                    type: string
                    description: User email
                    example: "irene@example.com"
                  picture:
                    type: string
                    description: Profile picture
                    example: null
                example: {'id': '1','name': 'Irene','surname': 'Sánchez', 'email': 'irene@gmail.com', 'picture': 'null'}
        "401":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"Password is wrong"}
...
  /pri/getEtapasRealizadas:
    get:
      tags:
      - "camino"
      summary: "Devuelve las etapas realizadas por un un usuario en un camino concreto"     
      parameters:
        - name: camino
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: "Returns an array of etapas"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: "object"
                example: [
    {
        "id": "988",
        "km": "24.2",
        "description": "...",
        "start": "Saint Jean Pied de Port",
        "finish": "Roncesvalles"
    },
    {
        "id": "989",
        "km": "21.4",
        "description": "...",
        "start": "Roncesvalles",
        "finish": "Zubiri"
    },
    {
        "id": "990",
        "km": "20.4",
        "description": "...",
        "start": "Zubiri",
        "finish": "Pamplona"
    },
    {
        "id": "991",
        "km": "23.9",
        "description": "...",
        "start": "Pamplona",
        "finish": "Puente de la Reina"
    }
]
        "400":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"Incorrect data recived"}
  /pri/addActivePath:
    post:
      tags:
      - "camino"
      summary: "If there isn't an active path it adds the one given to the DB"     
      parameters:
        - name: camino
          in: query
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Active path added"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"success"}
        "400":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"Incorrect data recived"}

        "422":
          description: "Invalid input"
          content:
            application/json:
              schema:
                type: string
                example: {"message":"User already has an active path"}
...